+++ plugin/sensor/touch_grid.cc
@@ -427,7 +427,11 @@ void TouchGrid::Visualize(const mjModel* m, mjData* d, const mjvOption* opt,
         continue;
       }
       if (scn->ngeom >= scn->maxgeom) {
-        mj_warning(d, mjWARN_VGEOMFULL, scn->maxgeom);
+        if (!scn->status) {
+          mju_warning("Pre-allocated visual geom buffer is full. "
+                      "Increase maxgeom above %d.", scn->maxgeom);
+          scn->status = 1;
+        }
         mj_freeStack(d);
         return;
       } else {